+++ src/mbs/adapters/database/models/__init__.py
@@ -1,3 +1,5 @@
+from mbs.db import Base
+
 from .article import ArticleTag
 from .article import Article
 from .tag import Tag
