+++ src/mbs/adapters/database/models/file.py
@@ -1,22 +1,22 @@
 import uuid
 
-from sqlalchemy import Column, UUID, DateTime, func, ForeignKey
+from sqlalchemy import Column, UUID, DateTime, func, ForeignKey, VARCHAR, BOOLEAN
 from sqlalchemy.orm import relationship
 
 from mbs.db import Base
 
 
-class Like(Base):
+class File(Base):
     """
-    The Like model
+    The File model
 
     """
-    __tablename__ = "likes"
+    __tablename__ = "files"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    owner_id = Column(UUID(as_uuid=True), nullable=False)
-    article_id = Column(UUID(as_uuid=True), ForeignKey("articles.id"), nullable=False)
-    article = relationship("models.tables.article.Article", back_populates="likes")
+    filename = Column(VARCHAR(255), nullable=False)
+    content_type = Column(VARCHAR(255), nullable=False)
+    is_uploaded = Column(BOOLEAN(), default=False)
 
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
